% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\name{create_data_files}
\alias{create_data_files}
\title{Create data files (.dat and .pin) for use with multispecies model, Hydra.}
\usage{
create_data_files(listOfParameters)
}
\arguments{
\item{listOfParameters}{see \code{\link{setup_default_inputs}}}
}
\value{
Nothing
}
\description{
Initial data files have been lazily loaded. To view their contents please run the doc rmd file
}
\section{Output}{


Two data files are output. A ".dat" and a ".pin" file. These are required by the Hydra code to run. The name of the file is determined by the variable \code{outputFilename} in the
list \code{listOfParameters}

Parts of data that user may want to control

Effort Fishing = implied by use of assessment and exploitation rate, or historical

from 1_3_3 to 1_4. added hockey stick and segmented, 4th fleet, and 4th guild
NOTE 4TH FLEET HAS NO EFFORT. HARD CODED TO OVERWRITE FILE. BEWARE
}

