% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datpin_files.R
\name{create_datpin_files}
\alias{create_datpin_files}
\title{Create data files (.dat and .pin) for use with multispecies model, Hydra.}
\usage{
create_datpin_files(listOfParameters, dataList)
}
\arguments{
\item{listOfParameters}{see \code{\link{setup_default_inputs}}}

\item{dataList}{A data set comprising all necessary inputs. This data set can be the lazily loaded data set or it can be modified in any way
to represent a faux assessment. See \code{\link{hydraData}} for details}
}
\value{
updated list of lazily loaded data "hydraData"
}
\description{
Initial data files have been lazily loaded. To view their contents please run the \code{\link{create_documentation}} rmd file
}
\section{Output}{


Two data files are output. A ".dat" and a ".pin" file. These are required by the Hydra code to run. The name of the file is determined by the variable \code{outputFilename} in the
list \code{listOfParameters}. By default the files are saved in the current working directory. You can specify a differnt location in \code{listofParameters}.
See \code{\link{setup_default_inputs}}
}

