% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydradata.R
\docType{package}
\name{hydradata}
\alias{hydradata}
\alias{hydradata-package}
\title{mshydradata: Manipulation of input data used in Hydra}
\description{
Tools to aid in quickly changing and documenting input parameters to Hydra
}
\details{
\itemize{
\item Manipulate input data to model
\item create documentation based on input data
\item creat .dat and .pin files used in running the multisecies model "Hydra"
}

To learn more about using \code{mshydradata}, start with the vignette: \code{browseVignettes(package="mshydradata")} or click \code{index} link below
}
