% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mshydradata.R
\docType{package}
\name{mshydradata}
\alias{mshydradata}
\alias{mshydradata-package}
\title{mshydradata: easy manipulation of input data used in Hydra}
\description{
An implementation of the LeMANS model as described in the Hall et al (2006) and Rochet et al (2011) publications.
}
\details{
\itemize{
\item Manipulate input data to model
\item create documentation based on input data
\item creat .dat and .pin files used in running the multisecies model "Hydra"
}

To learn more about using \code{mshydradata}, start with the vignette: \code{browseVignettes(package="mshydradata")}
}
