% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hcr.R
\name{set_hcr}
\alias{set_hcr}
\title{sets up Harvest Control Rules (HCR)}
\usage{
set_hcr(currentData = hydradata::hydraData, Nthresholds = 5,
  minMaxThresholds = c(0.1, 0.4), minMaxExploitations = c(0.05, 0.4),
  increment = 0.05)
}
\arguments{
\item{currentData}{List: The current hydraData in which changes need to be made}

\item{Nthresholds}{Numeric scalar: Number of thresholds in the step function}

\item{minMaxThresholds}{Numericric vector (1x2): Specifying the minimum and maximum threshold levels (proportion of B0) for management}

\item{minMaxExploitations}{Numeric vector (1x2): Specifying the minimum and maximum exploitation levels at \code{minMaxThresholds}}

\item{increment}{Numeric Scalar: increment to evaluate the number of scenarios between \code{minMaxExploitations}}
}
\value{
A list identical to \code{\link{hydraDataList}} with the fields below updated
}
\description{
Harvest control rules are evaluated using either a step function or a continuous linear function. For a continuous function all that is required
is \code{minMaxThreshold} and \code{minMaxExploitation}. For a step function, \code{Nthresholds}, \code{thresholds}, \code{exploitationOptions} are also required.
This function with create the step function \code{thresholds} and \code{exploitationOptions} given \code{minMaxThreshold}, \code{minMaxThreshold} and \code{Nthresholds}
}
\section{Updated fields}{


\code{exploitationOptions} - Numeric matrix: (\code{Nthresholds} x NScenarios). Range of possible scenarios for harvest control rule

\code{thresholds} - Numeric vector (1x \code{Nthresholds}): Specifying the threshold levels for which a change in exploitation will occur

\code{Nthresholds} - Numeric scalar: Specifying the number of thesholds

\code{minMaxThresholds} - Numeric vector (1x2): Specifying the minimum and maximum threshold levels (proportion of B0) for management
}

\section{Usage}{


Run this function to change the default set of step function scenarios.
}

