% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_documentation.R
\name{create_documentation}
\alias{create_documentation}
\title{Renders .rmd file to display all model inputs}
\usage{
create_documentation(outFile = "shit.html", outDir = getwd())
}
\description{
Renders .rmd file to display all model inputs
}
