% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraData.R
\docType{data}
\name{hydraData}
\alias{hydraData}
\title{Data for Hydra data nd pin files (mshydradata)}
\format{A list containing 133 elements
\describe{
    \item{speciesList}{Character vector: names of species in model}
    \item{}{}
    \item{}{}
}}
\source{
Gaichas et al. + publication (currently being written up)
}
\usage{
hydraData
}
\description{
All the required information for creating data and pin files are stored in this this. This
data is used to create documentation and to create dat and fin files for use in simulaion with Hydra
}
\seealso{
To visualize this data please run \code{create_documentation}
}
\keyword{datasets}
